Use WTForms for login form validation
Previously some validation was done by WTForms (using form validation error messaging) and some in the views (using flash messaging for errors).

This changes things so all form validation in the login/forgotten-password pages use WTForms validation.

Flash messaging is still used for confirmation and error messages not related to form validation.
