--- app/main/forms/login_form.py
@@ -1,14 +1,0 @@
-from flask.ext.wtf import Form
-from wtforms import StringField, PasswordField
-from wtforms.validators import DataRequired, Email
-
-
-class LoginForm(Form):
-    email_address = StringField('Email address', validators=[
-        DataRequired(message="Email can not be empty"),
-        Email(message="Please enter a valid email address")
-    ])
-
-    password = PasswordField('Password', validators=[
-        DataRequired(message="Please enter your password")
-    ])
