+++ app/templates/auth/change-password.html
@@ -1,27 +1,6 @@
 {% extends "_base_page.html" %}
 
 {% block main_content %}
-{% with messages = get_flashed_messages(with_categories=true) %}
-{% if messages %}
-    {% for category, message in messages %}
-        {% if category == 'error' %}
-        <div class="banner-destructive-without-action">
-        {% else %}
-        <div class="banner-success-without-action">
-        {% endif %}
-        <p class="banner-message">
-            {% if message == 'passwords_empty' %}
-            Please enter a new password.
-            {% endif %}
-            
-            {% if message == 'passwords_differ' %}
-            The passwords you entered do not match.
-            {% endif %}
-        </p>
-    {% endfor %}
-    </div>
-{% endif %}
-{% endwith %}
 
 <header class="page-heading">
     <h1>Change password</h1>
@@ -29,27 +8,60 @@
 
 <form action="{{ url_for('.update_password') }}" method="POST">
 
-    <input type="hidden" name="user-id" value="{{ user_id }}">
-    <input type="hidden" name="email-address" value="{{ email }}">
-    <div class="question">
-        <label for="password" class="question-heading-with-hint">
-            New password {% if email %}for {{ email }}{% endif %}
-        </label>
+    {% if form.errors %}
+        <div class="validation-masthead" aria-labelledby="validation-masthead-heading">
+            <h3 class="validation-masthead-heading" id="validation-masthead-heading">
+                There was a problem with your answer to the following questions
+            </h3>
+            {% for field_name, field_errors in form.errors|dictsort if field_errors %}
+            {% for error in field_errors %}
+            <a href="#example-textbox" class="validation-masthead-link">{{ form[field_name].label }}</a>
+            {% endfor %}
+            {% endfor %}
+        </div>
+    {% endif %}
+    
+    {{ form.hidden_tag() }}
 
-        <input type="password" name="password" id="password" class="text-box" value=""/>
-    </div>
+    {% if form.password.errors %}
+        <div class="validation-wrapper">
+    {% endif %}
+        <div class="question">
+            {{ form.password.label(class="question-heading-with-hint") }}
+            {% if form.email_address %}
+            <p class="hint">
+                New password for {{ form.email_address.data }} 
+            </p>
+            {% endif %}
+            {% if form.password.errors %}
+            <p class="validation-message" id="error-password-textbox">
+                {% for error in form.password.errors %}{{ error }}{% endfor %}
+            </p>
+            {% endif %}
+            {{ form.password(class="text-box") }}
+        </div>
+    {% if form.password.errors %}
+        </div>
+    {% endif %}
 
-    <div class="question">
-        <label for="confirm-password" class="question-heading">
-            Confirm password
-        </label>
-
-        <p class="hint">
-            Repeat password used above
-        </p>
-        <input type="password" name="confirm-password" id="confirm-password" class="text-box" value=""/>
-    </div>
-
+    {% if form.confirm_password.errors %}
+        <div class="validation-wrapper">
+    {% endif %}    
+        <div class="question">
+            {{ form.confirm_password.label(class="question-heading-with-hint") }}
+            <p class="hint">
+                Repeat password used above
+            </p>
+            {% if form.confirm_password.errors %}
+            <p class="validation-message" id="error-confirm-password-textbox">
+                {% for error in form.confirm_password.errors %}{{ error }}{% endfor %}
+            </p>
+            {% endif %}
+            {{ form.confirm_password(class="text-box") }}
+        </div>
+    {% if form.confirm_password.errors %}
+        </div>
+    {% endif %}
 
     <button class="button-save">Change password</button>
 </form>
