+++ app/templates/auth/forgotten-password.html
@@ -1,40 +1,40 @@
 {% extends "_base_page.html" %}
 
 {% block main_content %}
-{% with messages = get_flashed_messages(with_categories=true) %}
-{% if messages %}
-    {% for category, message in messages %}
-        {% if category == 'error' %}
-            <div class="banner-destructive-without-action">
-        {% else %}
-            <div class="banner-success-without-action">
-        {% endif %}
-            <p class="banner-message">
-                {% if message == 'email_sent' %}
-                If that Digital Marketplace supplier account exists, you will
-                be sent an email containing a link to reset your password.
-                {% endif %}
-                
-                {% if message == 'token_expired' %}
-                The token supplied has expired. Password reset links are only
-                valid for 24 hours. You can generate a new one using the form
-                below.
-                {% endif %}
-
-                {% if message == 'token_invalid' %}
-                The token supplied was invalid. You can generate a new one
-                using the form below.
-                {% endif %}
-            </p>
-    {% endfor %}
-            </div>
-{% endif %}
-{% endwith %}
-
 <header class="page-heading">
     <h1>Forgotten password</h1>
 </header>
 
+{% with messages = get_flashed_messages(with_categories=true) %}
+    {% if messages %}
+        {% for category, message in messages %}
+            {% if category == 'error' %}
+                <div class="banner-destructive-without-action">
+            {% else %}
+                <div class="banner-success-without-action">
+            {% endif %}
+                    <p class="banner-message">
+                        {% if message == 'email_sent' %}
+                        If that Digital Marketplace supplier account exists, you will
+                        be sent an email containing a link to reset your password.
+                        {% endif %}
+            
+                        {% if message == 'token_expired' %}
+                        The token supplied has expired. Password reset links are only
+                        valid for 24 hours. You can generate a new one using the form
+                        below.
+                        {% endif %}
+            
+                        {% if message == 'token_invalid' %}
+                        The token supplied was invalid. You can generate a new one
+                        using the form below.
+                        {% endif %}
+                    </p>
+                </div>
+        {% endfor %}
+    {% endif %}
+{% endwith %}
+    
 <form action="{{ url_for('.forgotten_password') }}" method="POST" id="forgotPasswordForm">
 
     {% if form.errors %}
@@ -42,13 +42,11 @@
         <h3 class="validation-masthead-heading" id="validation-masthead-heading">
             There was a problem with your answer to the following questions
         </h3>
-
         {% for field_name, field_errors in form.errors|dictsort if field_errors %}
         {% for error in field_errors %}
         <a href="#example-textbox" class="validation-masthead-link">{{ form[field_name].label }}</a>
         {% endfor %}
         {% endfor %}
-
     </div>
     {% endif %}
     
@@ -57,22 +55,18 @@
     {% if form.email_address.errors %}
         <div class="validation-wrapper">
     {% endif %}
-            
         <div class="question">
-            <label for="email_address" class="question-heading-with-hint">
-                Email address
-            </label>
-            {% if form.email_address.errors %}
-            <p class="validation-message" id="error-example-textbox">
-                {% for error in form.email_address.errors %}{{ error }}{% endfor %}
-            </p>
-            {% endif %}
+            {{ form.email_address.label(class="question-heading-with-hint") }}
             <p class="hint">
                 This is the email address used to register with the digital marketplace.
             </p>
-            <input type="text" name="email_address" id="email_address" class="text-box" value=""/>
+            {% if form.email_address.errors %}
+            <p class="validation-message" id="error-email-address-textbox">
+                {% for error in form.email_address.errors %}{{ error }}{% endfor %}
+            </p>
+            {% endif %}
+            {{ form.email_address(class="text-box") }}
         </div>
-            
     {% if form.email_address.errors %}
         </div>
     {% endif %}
