+++ app/templates/auth/login.html
@@ -4,67 +4,66 @@
 <header class="page-heading">
     <h1>Supplier login</h1>
 </header>
+
+{% with messages = get_flashed_messages(with_categories=true) %}
+    {% if messages %}
+        {% for category, message in messages %}
+            {% if category == 'error' %}
+                <div class="banner-destructive-without-action">
+            {% else %}
+                <div class="banner-success-without-action">
+            {% endif %}
+                    <p class="banner-message">
+                        {% if message == 'no_account' %}
+                        Sorry, we couldn't find a supplier account with that username
+                        and password.
+                        {% endif %}
+            
+                        {% if message == 'password_updated' %}
+                        You have successfully changed your password.
+                        {% endif %}
+            
+                        {% if message == 'password_not_updated' %}
+                        Could not update password due to an error.
+                        {% endif %}
+                    </p>
+                </div>
+        {% endfor %}
+    {% endif %}
+{% endwith %}
+
 <form action="{{ url_for('.process_login') }}" method="POST">
 
     {% if form.errors %}
-    <div class="validation-masthead" aria-labelledby="validation-masthead-heading">
-        <h3 class="validation-masthead-heading" id="validation-masthead-heading">
-            There was a problem with your answer to the following questions
-        </h3>
-
-        {% for field_name, field_errors in form.errors|dictsort if field_errors %}
-        {% for error in field_errors %}
-        <a href="#example-textbox" class="validation-masthead-link">{{ form[field_name].label }}</a>
-        {% endfor %}
-        {% endfor %}
-
-    </div>
-    {% endif %}
-
-    {% with messages = get_flashed_messages(with_categories=true) %}
-        {% if messages %}
-            {% for category, message in messages %}
-                {% if category == 'error' %}
-                    <div class="banner-destructive-without-action">
-                {% else %}
-                    <div class="banner-success-without-action">
-                {% endif %}
-            <p class="banner-message">
-                {% if message == 'no_account' %}
-                Sorry, we couldn't find a supplier account with that username 
-                and password.
-                {% endif %}
-                
-                {% if message == 'password_updated' %}
-                You have successfully changed your password.
-                {% endif %}
-                
-                {% if message == 'password_not_updated' %}
-                Could not update password due to an error.
-                {% endif %}
-            </p>
+        <div class="validation-masthead" aria-labelledby="validation-masthead-heading">
+            <h3 class="validation-masthead-heading" id="validation-masthead-heading">
+                There was a problem with your answer to the following questions
+            </h3>
+            {% for field_name, field_errors in form.errors|dictsort if field_errors %}
+            {% for error in field_errors %}
+            <a href="#example-textbox" class="validation-masthead-link">{{ form[field_name].label }}</a>
+            {% endfor %}
             {% endfor %}
         </div>
-        {% endif %}
-     {% endwith %}
+    {% endif %}
 
     {{ form.hidden_tag() }}
 
     {% if form.email_address.errors %}
         <div class="validation-wrapper">
     {% endif %}
-    <div class="question">
-        {{ form.email_address.label(class="question-heading-with-hint") }}
-        <p class="hint">
-            Email used to register with the digital marketplace.
-        </p>
-        {% if form.email_address.errors %}
-        <p class="validation-message" id="error-email-address-textbox">
-            {% for error in form.email_address.errors %}{{ error }}{% endfor %}
-        </p>
-        {% endif %}
-        {{ form.email_address(class="text-box") }}
-    </div>
+        <div class="question">
+            {{ form.email_address.label(class="question-heading-with-hint") }}
+            <p class="hint">
+                Email used to register with the digital marketplace.
+            </p>
+            {% if form.email_address.errors %}
+            <p class="validation-message" id="error-email-address-textbox">
+                {% for error in form.email_address.errors %}{{ error }}{% endfor %}
+            </p>
+            {% endif %}
+            {{ form.email_address(class="text-box") }}
+        </div>
     {% if form.email_address.errors %}
         </div>
     {% endif %}
@@ -72,20 +71,20 @@
     {% if form.password.errors %}
         <div class="validation-wrapper">
     {% endif %}                        
-    <div class="question">
-        {{ form.password.label(class="question-heading") }}
-        {% if form.password.errors %}
-        <p class="validation-message" id="error-password-textbox">
-            {% for error in form.password.errors %}{{ error }}{% endfor %}
-        </p>
-        {% endif %}
-        {{ form.password(class="text-box") }}
-    </div>
+        <div class="question">
+            {{ form.password.label(class="question-heading") }}
+            {% if form.password.errors %}
+            <p class="validation-message" id="error-password-textbox">
+                {% for error in form.password.errors %}{{ error }}{% endfor %}
+            </p>
+            {% endif %}
+            {{ form.password(class="text-box") }}
+        </div>
     {% if form.password.errors %}
         </div>
     {% endif %}
 
-                <button class="button-save">Log in</button>
+    <button class="button-save">Log in</button>
     <header class="page-heading-smaller">
         <p class="context">
             <a href="{{ url_for('.forgotten_password') }}">Forgotten password</a>
